app/pyproject.toml
[tool.poetry]
name = "bet-maker"
version = "0.1.0"
description = ""
authors = ["Mike <1651319@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
python-dotenv = "^1.0.1"
databases = {extras = ["asyncpg"], version = "^0.9.0"}
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"